Move all code to src/pointnet
