+++ src/pointnet/part_seg/test.py
@@ -7,7 +7,6 @@ import sys
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(BASE_DIR)
 sys.path.append(os.path.dirname(BASE_DIR))
-import provider
 import pointnet_part_seg as model
 
 parser = argparse.ArgumentParser()
