+++ src/pointnet/part_seg/testing_ply_file_list.txt
