+++ src/pointnet/sem_seg/collect_indoor3d_data.py
