+++ src/pointnet/sem_seg/download_data.sh
