+++ src/pointnet/sem_seg/gen_indoor3d_h5.py
