+++ src/pointnet/sem_seg/meta/all_data_label.txt
