+++ src/pointnet/sem_seg/meta/anno_paths.txt
