+++ src/pointnet/sem_seg/meta/area6_data_label.txt
