+++ src/pointnet/train.py
@@ -1,6 +1,4 @@
 import argparse
-import math
-import h5py
 import numpy as np
 import tensorflow as tf
 import socket
@@ -11,8 +9,7 @@ BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(BASE_DIR)
 sys.path.append(os.path.join(BASE_DIR, 'models'))
 sys.path.append(os.path.join(BASE_DIR, 'utils'))
-import provider
-import tf_util
+from pointnet import provider
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--gpu', type=int, default=0, help='GPU to use [default: GPU 0]')
